+++ src/client/rest/RESTMethods.js
@@ -1,81 +1,10 @@
 const Constants = require('../../util/Constants');
 const Endpoints = Constants.Endpoints;
-const Util = require('../../util/Util');
-
-const User = require('../../structures/User');
-const GuildMember = require('../../structures/GuildMember');
 
 class RESTMethods {
   constructor(restManager) {
     this.rest = restManager;
     this.client = restManager.client;
-  }
-
-  sendMessage(channel, content, { tts, nonce, embed, disableEveryone, split, code, reply } = {}, files = null) {
-    return new Promise((resolve, reject) => { // eslint-disable-line complexity
-      if (typeof content !== 'undefined') content = this.client.resolver.resolveString(content);
-
-      // The nonce has to be a uint64 :<
-      if (typeof nonce !== 'undefined') {
-        nonce = parseInt(nonce);
-        if (isNaN(nonce) || nonce < 0) throw new RangeError('Message nonce must fit in an unsigned 64-bit integer.');
-      }
-
-      if (content) {
-        if (split && typeof split !== 'object') split = {};
-
-        // Wrap everything in a code block
-        if (typeof code !== 'undefined' && (typeof code !== 'boolean' || code === true)) {
-          content = Util.escapeMarkdown(this.client.resolver.resolveString(content), true);
-          content = `\`\`\`${typeof code !== 'boolean' ? code || '' : ''}\n${content}\n\`\`\``;
-          if (split) {
-            split.prepend = `\`\`\`${typeof code !== 'boolean' ? code || '' : ''}\n`;
-            split.append = '\n```';
-          }
-        }
-
-        // Add zero-width spaces to @everyone/@here
-        if (disableEveryone || (typeof disableEveryone === 'undefined' && this.client.options.disableEveryone)) {
-          content = content.replace(/@(everyone|here)/g, '@\u200b$1');
-        }
-
-        // Add the reply prefix
-        if (reply && !(channel instanceof User || channel instanceof GuildMember) && channel.type !== 'dm') {
-          const id = this.client.resolver.resolveUserID(reply);
-          const mention = `<@${reply instanceof GuildMember && reply.nickname ? '!' : ''}${id}>`;
-          content = `${mention}${content ? `, ${content}` : ''}`;
-          if (split) split.prepend = `${mention}, ${split.prepend || ''}`;
-        }
-
-        // Split the content
-        if (split) content = Util.splitMessage(content, split);
-      } else if (reply && !(channel instanceof User || channel instanceof GuildMember) && channel.type !== 'dm') {
-        const id = this.client.resolver.resolveUserID(reply);
-        content = `<@${reply instanceof GuildMember && reply.nickname ? '!' : ''}${id}>`;
-      }
-
-      const send = chan => {
-        if (content instanceof Array) {
-          const messages = [];
-          (function sendChunk(list, index) {
-            const options = index === list.length ? { tts, embed } : { tts };
-            chan.send(list[index], options, index === list.length ? files : null).then(message => {
-              messages.push(message);
-              if (index >= list.length - 1) return resolve(messages);
-              return sendChunk(list, ++index);
-            });
-          }(content, 0));
-        } else {
-          this.rest.request('post', Endpoints.Channel(chan).messages, {
-            data: { content, tts, nonce, embed },
-            files,
-          }).then(data => resolve(this.client.actions.MessageCreate.handle(data).message), reject);
-        }
-      };
-
-      if (channel instanceof User || channel instanceof GuildMember) this.createDM(channel).then(send, reject);
-      else send(channel);
-    });
   }
 
   sendWebhookMessage(webhook, content, { avatarURL, tts, disableEveryone, embeds, username } = {}, file = null) {
