+++ src/structures/shared/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  search: require('./Search'),
+  sendMessage: require('./SendMessage'),
+};
